refactor: auth plugin gf-13
