+++ apps/backend/src/libs/modules/token/base-token.module.ts
@@ -1,4 +1,4 @@
-import { decodeJwt, type JWTPayload, jwtVerify, SignJWT } from "jose";
+import { type JWTPayload, jwtVerify, SignJWT } from "jose";
 
 import { type Token } from "./libs/types/types.js";
 
@@ -25,10 +25,6 @@ class BaseToken implements Token {
 			.setIssuedAt()
 			.setExpirationTime(this.expirationTime)
 			.sign(this.secret);
-	}
-
-	public decodeToken(token: string): JWTPayload {
-		return decodeJwt(token);
 	}
 
 	public async verifyToken(token: string): Promise<JWTPayload> {
