+++ apps/backend/src/libs/modules/token/libs/types/token.type.ts
@@ -2,7 +2,6 @@ import { type JWTPayload } from "jose";
 
 type Token = {
 	createToken: (payload: JWTPayload) => Promise<string>;
-	decodeToken: (token: string) => JWTPayload;
 	verifyToken: (token: string) => Promise<JWTPayload>;
 };
