+++ apps/backend/src/libs/plugins/libs/helpers/helpers.ts
@@ -1,0 +1,1 @@
+export { isTokenExpiredError } from "./token-expiration.js";
