+++ apps/backend/src/libs/plugins/libs/helpers/token-expiration.ts
@@ -1,0 +1,6 @@
+function isTokenExpiredError(error: Error): boolean {
+	// eslint-disable-next-line quotes
+	return error.message.includes('"exp" claim timestamp check failed');
+}
+
+export { isTokenExpiredError };
