+++ packages/shared/src/libs/enums/exception-message.enum.ts
@@ -2,7 +2,8 @@ const ExceptionMessage = {
 	EMAIL_USED: "Email address is already in use.",
 	INVALID_CREDENTIALS: "Invalid credentials.",
 	INVALID_TOKEN: "Invalid token",
-	NO_TOKEN_PROVIDED: "No token provided",
+	NO_TOKEN_PROVIDED: "No token provided.",
+	TOKEN_EXPIRED: "Token is expired.",
 	USER_NOT_FOUND: "User not found.",
 } as const;
